Initial Accounts refactor
